порвал файл с тестами. решарпер с ума сходил на большом файле
